Refactor the builtins to use more overloads (#1751)
This is a wholesale refactor of the builtins, splitting them into 3
files for generic, math + integer, as well as relational builtins.
As part of the refactoring, some functions have been refactored from
solely using templates to using overloads and templates, while others
have had their templates updated according to the latest spec.

Because of the support for floats and vectors/marrays, the actual
`builtin_interface` has been changed, particularly with the functions
taking pointer arguments.