[CBS] Insert PHIs manually if required. (#1909)
* Add test case from issue #1898.

* [CBS] Insert PHIs manually if required.

When splitting a loop in CBS, we might get "instruction not dominated" issues.
Most of them were already captured by the existing infrastructure.

However, the issue with #1898, was that an inner loop was split, and the outer loop wasn't split.
All loop indices are uniform in the outer loop.
Therefore, they were loaded only once in the uniload block.
However, when the outer loop takes its backedge (after the inner loop exited, thus without another barrier), the outer loop's index is not updated for the first iteration of the inner loop.
Therefore, we should create a PHI node after the backedge of the outer loop to select between the value loaded in the uniload block and the value from the backedge of the outer loop.

* Restore compatibility with older LLVMs.

* [CBS] Use dominance frontier to find all the spots where a PHI might be needed.

A little intro on DFs and why they're interesing for this task:
https://pages.cs.wisc.edu/~fischer/cs701.f05/lectures/Lecture25.4up.pdf

This is a more fundamental approach to the issue, so I hope it'll hold up against whatever we may throw against it.

At the moment, this is an overapproximation, i.e. there might be potential to prevent creating too many PHIs, but I expect them to be optimized away, as they're just dead code.

---------

Co-authored-by: Aksel Alpay <aksel.alpay@uni-heidelberg.de>